Add more options to ctx
