+++ crescent/context.py
@@ -1,22 +1,27 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, cast
+from typing import TYPE_CHECKING, Dict, Mapping, cast
 
 from attr import define
 from hikari import (
     UNDEFINED,
     CacheAware,
+    CommandInteractionOption,
+    CommandType,
     Guild,
     GuildChannel,
     Member,
     MessageFlag,
+    OptionType,
     ResponseType,
     RESTAware,
     Snowflake,
     User,
 )
+import hikari
 
-from crescent.utils import map_or
+from crescent.utils import map_or, unwrap
+from crescent.mentionable import Mentionable
 
 if TYPE_CHECKING:
     from typing import Any, Optional, Sequence, Type
@@ -35,12 +40,60 @@ if TYPE_CHECKING:
     )
     from hikari.api import ComponentBuilder
 
+    from crescent.typedefs import OptionTypesT
+
 
 class RestAndCacheAware(RESTAware, CacheAware):
     ...
 
 
 __all__: Sequence[str] = ("Context",)
+
+
+_VALUE_TYPE_LINK: Dict[OptionType | int, str] = {
+    OptionType.ROLE: "roles",
+    OptionType.USER: "users",
+    OptionType.CHANNEL: "channels",
+}
+
+
+def _options_to_kwargs(
+    interaction: CommandInteraction,
+    options: Optional[Sequence[CommandInteractionOption]],
+) -> Dict[str, Any]:
+    if not options:
+        return {}
+
+    return {option.name: _extract_value(option, interaction) for option in options}
+
+
+def _extract_value(option: CommandInteractionOption, interaction: CommandInteraction) -> Any:
+    if option.type is OptionType.MENTIONABLE:
+        return Mentionable._from_interaction(interaction)
+
+    resolved_type: Optional[str] = _VALUE_TYPE_LINK.get(option.type)
+
+    if resolved_type is None:
+        return option.value
+
+    resolved = getattr(interaction.resolved, resolved_type)
+    return resolved[option.value]
+
+
+def _resolved_data_to_kwargs(interaction: CommandInteraction) -> Dict[str, Message | User]:
+    if not interaction.resolved:
+        raise ValueError("interaction.resoved should be defined when running this function")
+
+    if interaction.resolved.messages:
+        return {"message": next(iter(interaction.resolved.messages.values()))}
+    if interaction.resolved.members:
+        return {"user": next(iter(interaction.resolved.members.values()))}
+    if interaction.resolved.users:
+        return {"user": next(iter(interaction.resolved.users.values()))}
+
+    raise AttributeError(
+        "interaction.resolved did not have property `messages`, `members`, or `users`"
+    )
 
 
 @define
@@ -59,11 +112,39 @@ class Context:
     user: User
     member: Optional[Member]
 
+    command: str
+    command_type: hikari.CommandType
+    group: Optional[str]
+    sub_group: Optional[str]
+    options: Dict[str, Any]
+
     _has_replied: bool = False
     _used_first_resp: bool = False
 
     @classmethod
     def _from_command_interaction(cls: Type[Context], interaction: CommandInteraction) -> Context:
+        name: str = interaction.command_name
+        group: Optional[str] = None
+        sub_group: Optional[str] = None
+        options = interaction.options
+
+        if options:
+            option = options[0]
+            if option.type == 1:
+                group = name
+                name = option.name
+                options = option.options
+            elif option.type == 2:
+                group = interaction.command_name
+                sub_group = option.name
+                name = unwrap(option.options)[0].name
+
+        callback_options: Mapping[str, OptionTypesT | Message | User]
+        if interaction.command_type is CommandType.SLASH:
+            callback_options = _options_to_kwargs(interaction, options)
+        else:
+            callback_options = _resolved_data_to_kwargs(interaction)
+
         return cls(
             app=cast(RestAndCacheAware, interaction.app),
             application_id=interaction.application_id,
@@ -75,6 +156,11 @@ class Context:
             guild_id=interaction.guild_id,
             user=interaction.user,
             member=interaction.member,
+            command=name,
+            group=group,
+            sub_group=sub_group,
+            command_type=hikari.CommandType(interaction.command_type),
+            options=callback_options,
         )
 
     @property
