+++ examples/hooks/hooks.py
@@ -1,12 +1,7 @@
 import crescent
 
 
-async def first_hook(
-    ctx: crescent.Context,
-    # options is a shallow copy of the options from the interaction
-    options: crescent.CommandOptionsT,
-    # You can also return `None`
-) -> crescent.HookResult:
+async def first_hook(ctx: crescent.Context) -> crescent.HookResult:  # you can also return None
     print("Here first.")
     # Setting exit to true prevents any following hooks and the callback from
     # being run.
@@ -15,12 +10,9 @@ async def first_hook(
 
 
 def second_hook(number: int) -> crescent.HookCallbackT:
-    async def inner(
-        ctx: crescent.Context, options: crescent.CommandOptionsT
-    ) -> crescent.HookResult:
+    async def inner(ctx: crescent.Context) -> None:
         print(f"Here second. Number is {number}")
-        options["number"] = number
-        return crescent.HookResult(options=options)
+        ctx.options["numeber"] = number
 
     return inner
